rag-service/requirements.txt
fastapi
uvicorn
langchain
faiss-cpu
ollama
redis
requests
python-dotenv